+++ CoCowork.DB/StoredProcedures/MiniOffice/MiniOffice_Update.sql
@@ -2,7 +2,7 @@
 	@Id int,
 	@Name varchar(30),
 	@AmountOfPlaces int,
-	@PricePerDay int,
+	@PricePerDay decimal(10,2),
 	@IsActive bit
 AS
 BEGIN
