+++ CoCowork.DB/StoredProcedures/Place/Place_Insert.sql
@@ -1,7 +1,7 @@
 ﻿CREATE PROC dbo.Place_Insert
 	@MiniOfficeId int,
-	@PricePerDay int,
-	@PriceFixedPerDay int
+	@PricePerDay decimal(10,2),
+	@PriceFixedPerDay decimal(10,2)
 AS
 BEGIN
 	insert into dbo.Place
